Pipes work and tested, still get clogged need to use dt for water pump. Barrels are fluid container now and amount is a float

+++ Fluids/FluidContainer.cs
@@ -15,7 +15,7 @@ namespace Base_Building_Game
     {
         public interface FluidContainer
         {
-            public int amount { get; set; }
+            public float amount { get; set; }
 
             public short type { get; set; }
